requirements.txt
docutils>=0.15.2
kivy-deps.glew>=0.1.12
kivy-deps.sdl2>=0.1.23
Kivy-Garden>=0.1.4
kivymd>=0.104.0
opencv-contrib-python>=4.2.0.34
Pygments>=2.5.2
pypiwin32>=223
pywin32>=227
numpy>=1.18.4
